Fit series of linear regression models to multiple responses with shared design matrix

Fit regression model Y[,j] ~ X_design for each feature j

Usage

lmFitResponses(Y, X, ids, Weights, detail = 0L, nthreads = 1L)

Arguments

Y

matrix of responses as columns

X

design matrix

ids

vector<string> storing identifier for each column in Y

Weights

matrix sample-level weights the same dimension as Y

detail

level of model detail returned, with LOW = 0, MEDIUM = 1, HIGH = 2. LOW (beta, se, sigSq, rdf), MEDIUM (vcov), HIGH (residuals), MOST (hatvalues)

nthreads

number of threads. Each model is fit in serial, analysis is parallelized across responses.

Value

List of parameter estimates with entries coef, se, sigSq, rdf and other depending on detail

Details

Since the weights vary for each response, each model is computed separately without recycling precomputed values

Examples

n = 100
m = 5
nc = 2
set.seed(1)
Y = matrix(rnorm(n*m), n, m)
X = matrix(rnorm(n*nc), n,nc)
colnames(Y) = seq(m)
W = matrix(runif(n*m), n,m) 

# fit regressions with model j using Y[,j] as a response
fit = lmFitResponses(Y, X, colnames(Y), W)

# examine results
lapply(fit, head, 2)
#> $coef
#>          [,1]        [,2]
#> 1 -0.05603001  0.04668949
#> 2 -0.09457993 -0.04716347
#> 
#> $se
#>         [,1]       [,2]
#> 1 0.09124254 0.08104439
#> 2 0.10212583 0.08152607
#> 
#> $sigSq
#>         1         2 
#> 0.3532643 0.4573510 
#> 
#> $rdf
#>  1  2 
#> 98 98 
#>